adding costed and turning angle terms, starting to tune and debug the 10% nonconveragance 0.00e0 issues

+++ smac_planner/include/smac_planner/minimal_costmap.hpp
@@ -66,7 +66,7 @@ public:
    */
   inline double getCost(const int & mx, const int & my) const
   {
-    return _char_costmap[getIndex(mx, my)];
+    return static_cast<double>(_char_costmap[getIndex(mx, my)]);
   }
 
   /**
